added downhillsimplex method, seems to not work with multiple clusters though

+++ pluginsimpleforest/qsm/algorithm/spherefollowing/sf_spherefollowingParameters.h
@@ -34,11 +34,11 @@
 #include <pcl/segmentation/sac_segmentation.h>
 
 struct SF_SphereFollowingOptimizationParameters {
-  float _euclideanClusteringDistance = 0.03f;
-  float _sphereRadiusMultiplier = 2.0f;
-  float _epsilonSphere = 0.035f;
-  float _minRadius = 0.07f;
-  float _medianRadiusMultiplier = 3.0f;
+  double _euclideanClusteringDistance = 0.03f;
+  double _sphereRadiusMultiplier = 2.0f;
+  double _epsilonSphere = 0.035f;
+  double _minRadius = 0.07f;
+  double _medianRadiusMultiplier = 3.0f;
   SF_SphereFollowingOptimizationParameters() {}
   SF_SphereFollowingOptimizationParameters(float euclideanClusteringDistance,
                                            float sphereRadiusMultiplier,
